refactor: renamed auth context  to auth provider
