+++ src/renderer/src/pages/BrowserAuth.tsx
@@ -2,7 +2,7 @@ import Button from '@/components/button/Button'
 import { LogoIcon, SendIcon } from '@/components/icons'
 import { Input } from '@/components/input/input'
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/popover/Popover'
-import useAuth from '@/context/auth/useAuth'
+import useAuth from '@/providers/authProvider/useAuth'
 import { AUTH_API } from '@/services/auth'
 import { copyToClipboard } from '@/utils/helpers'
 import { FC, useEffect } from 'react'
