+++ src/renderer/src/providers/authProvider/authEnums.ts
